Modernize build for current GCC/libobjc toolchain
Fix bit-rot so a single "make depend; make; make install" builds and
installs all 7 libraries on a modern Debian/GCC 14 toolchain.

Build system:
- makedir/targetLinux: APPLY = $(MAKEDIR)/apply (absolute, was unresolved
  on PATH during make depend).
- makedir/common: install :: all, so install depends on the libraries
  being built (fixes "cannot stat libbclib.a" from double-colon ordering).
- Add in-tree -I paths to sibling modules' interfaces in every dependent
  module's Makefile (benlib, bolib, bplib, bpcombi, bprib, and bplib
  tests). gcc -M could not resolve cross-module headers
  (bclib/interfaces/names.h, BcString.h, ...), which produced empty
  Makefile.depend and was the real reason "make depend" was needed twice.
  One pass now suffices.

bplib (modern C++ rot under -Werror -pedantic-errors):
- Cast CARD32 indices through (size_t) before (void*) to silence
  -Werror=int-to-pointer-cast (BpArray.cc, BpList.cc, BpObject.cc).
- Space around FMT_CARD32 string macro for -Werror=literal-suffix
  (PrintOnLevel.cc).

bolib (legacy Objective-C Object root class):
- Modern libobjc's Object only declares -class/-isEqual:. Add an
  Object (BoCompat) category reimplementing the legacy methods bolib uses
  (+alloc, -init, -free, +name/-name, -error:, -perform:/with:/with:with:)
  on the GNU runtime (class_createInstance, object_dispose, objc_msg_lookup,
  class_getName/object_getClassName). Declared in BoObject.h, implemented
  in BoObject.m so the category is always linked from the static archive.

+++ bplib/sources/BpArray.cc
@@ -199,7 +199,7 @@ METHOD(BpArray,resize,(CARD32 newSize),BpArray*)
 METHOD(BpArray,objectAt,(CARD32 index),BpObject*)
 {
     if(index>=lsize){
-        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)index);
+        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)(size_t)index);
     }
     return(elements[index]);
 }//objectAt:;
@@ -211,7 +211,7 @@ METHOD(BpArray,replaceObjectAtWith,
     BpObject*           old;
             
     if(index>=lsize){
-        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)index);
+        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)(size_t)index);
     }
     old=elements[index];
     elements[index]=newObject;
@@ -231,7 +231,7 @@ METHOD(BpArray,insertObjectAt,(BpObject* newObject,CARD32 index),BpArray*)
     CARD32              i;
             
     if(index>=lsize){
-        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)index);
+        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)(size_t)index);
     }
     i=lsize-1;
     while(i>index){
@@ -253,7 +253,7 @@ METHOD(BpArray,removeObjectAt,(CARD32 index),BpObject*)
     BpObject*           old;
             
     if(index>=lsize){
-        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)index);
+        BcRAISE(BpArray_eBadIndex,(void*)this,(void*)(size_t)index);
     }
     old=elements[index];
     i=index;
